Fallback to IndieAuth.com
Fallback to indiieauth.com if indieauth plugin is not installed.

Right now the indieauth plugin is not working for me, so this attempts
to provide an alternative.

Getting an error still, though this gets much further. For now I have
added a debugging hard-coded user_id =1.

Channels are hard-coded for testing. They show up in monocle.p3k.io but
not in alltogethernow.io.

Still, making progress!