Improved meta for aggregated posts
Set title to {channel}|{permalink} which allows for the same feed to be subscribed in multiple channels

Set {post_date} and {post_modified} based on the {published} and {updated} properties. This makes it easier to sort the timeline chronologically

Thinking about how to handle the broken json in the content->html. I think the best option will be to store the json in a custom field instead of as as the post content. This will also avoid wordpress's restrictions about invalid html, which would probably cause problems with some posts