Improved parsing
Improved parsing

If a post is missing or has incomplete author info, supplement with author info from the feed itself. (The Parse_MF2 class did this to some extent before, but was missing some items)

Instead of fetching each post individually, aggregator now relies wholly on the h-feed itself. This is considerably faster, and there aren't many cases where the permalink has extra information that is meaningful.

Moved away from using mergeparse.  Accordingly, moved its cleaning tasks to a new function `clean_post`

Adds a variable indicating the type of a feed (rss or h-feed).  This was used while I was writing this comit, but may not be useful anymore.

Some tweaks to class-parse-mf2. It was throwing numerous errors when passed incomplete data - particularly malformed or empty 'properties' arrays.

Modified functions for encoding/decoding arrays. (encode_array and decode_array) Not sure these are really needed but I'll wait and see for now.