Fixes 'channels' json AND adds pagination to timeline
The channel list was being returned along with a list of each channel's subscribed feeds. This does not conform to the Microsub spec and may be causing indigenous for ios to have trouble loading channels (https://github.com/EdwardHinkle/indigenous-ios/issues/224)

Added pagination when fetching timeline.  Tested briefly and it seems to work so I'll cross my fingers.  Addresses #7