Add options for filtering channels
When new channels are created they are assigned a list of post_types to allow.
When new posts are created, they are assigned a taxonomy for post-type (based on parse-this's post type discovery)
When timeline is called, it will only show post types that are within this list.

A bit of work on admin settings page for managing these post type filters per channel, but not done yet.

Todo:
- finish admin UI
- handle 'other' posts in timeline
- allow all rss posts, since they don't have post-type-discovery